more something

+++ frontend/src/pages/add/editDialect/EditDialect.tsx
@@ -1,17 +1,21 @@
-import { Table } from '@radix-ui/themes';
+// @ts-ignore
+import { useTranslation } from 'react-i18next';
+import { AlertDialog, Button, Dialog, Flex, Table } from '@radix-ui/themes';
 import { motion } from 'framer-motion';
 import { useDialects } from '../../../hooks/useDialects';
 import { useEffect, useState } from 'react';
 import { IAllDialect } from '../../../app.interface';
-import { EyeOff, Pencil, Trash2 } from 'lucide-react';
-import { Link } from 'react-router-dom';
+import { Pencil, Trash2 } from 'lucide-react';
 import toast from 'react-hot-toast';
 import { DialectService } from '../../../services/dialect.service';
+import EditDialectPage from './editDialectPage/EditDialectPage';
+import AddDialect from '../addDialect/AddDialect';
+import ButtonUI from '../../../components/ui/button/Button';
 
 const EditDialect = () => {
-  const { data } = useDialects();
+  const { t, i18n } = useTranslation('translation');
 
-  console.log(data);
+  const { data } = useDialects();
 
   const [dialectsData, setDialectsData] = useState<IAllDialect[]>([]);
 
@@ -24,15 +28,25 @@ const EditDialect = () => {
   }, [data]);
 
   const deleteData = async (_id: string, title: string) => {
-    if (window.confirm(`Are you sure you want to delete ${title}?`)) {
-      try {
-        await DialectService.deleteDialect(_id)
-        toast.success('Успешно удалено!');
-        window.location.reload()
-      } catch (error) {
-        toast.error('Ошибка при удалении');
-      }
+    try {
+      await DialectService.deleteDialect(_id);
+      toast.success('Успешно удалено!');
+      window.location.reload();
+    } catch (error) {
+      toast.error('Ошибка при удалении');
     }
+  };
+
+  const getRegionValue = (item: any) => {
+    // Проверяем текущий язык
+    const currentLanguage = i18n.language;
+
+    // Если текущий язык - казахский, возвращаем item.kzRegion
+    if (currentLanguage === 'kz') {
+      return item.kzRegion;
+    }
+    // В противном случае возвращаем item.enRegion
+    return item.enRegion;
   };
 
   return (
@@ -41,21 +55,23 @@ const EditDialect = () => {
       animate={{ opacity: 1 }}
       transition={{ duration: 0.8 }}
       className='w-fit h-fit flex flex-col items-center justify-center gap-5 shadow bg-white dark:bg-transparent px-20 py-5 rounded-lg'
-      style={{scrollSnapAlign: 'center'}}
+      style={{ scrollSnapAlign: 'center' }}
     >
       <div>
         <Table.Root
           // variant='surface'
-          style={{ maxHeight: 400, minHeight: 200, overflowY: 'auto' }}
+          style={{ maxHeight: 400, overflowY: 'auto' }}
         >
           <Table.Header>
             <Table.Row>
-              <Table.ColumnHeaderCell width={350}>Title</Table.ColumnHeaderCell>
               <Table.ColumnHeaderCell width={350}>
-                Region
+                {t('control.dialect.title')}
+              </Table.ColumnHeaderCell>
+              <Table.ColumnHeaderCell width={350}>
+                {t('control.dialect.region')}
               </Table.ColumnHeaderCell>
               <Table.ColumnHeaderCell width={200}>
-                Action
+                {t('control.dialect.action')}
               </Table.ColumnHeaderCell>
             </Table.Row>
           </Table.Header>
@@ -65,19 +81,55 @@ const EditDialect = () => {
               dialectsData.map((item: any, index: number) => (
                 <Table.Row key={index}>
                   <Table.RowHeaderCell>{item.title}</Table.RowHeaderCell>
-                  <Table.Cell>{item.kzRegion}</Table.Cell>
+                  <Table.Cell>{getRegionValue(item)}</Table.Cell>
                   <Table.Cell>
                     <div className='flex items-center gap-5'>
-                      <Link to={`/add/${item._id}`}>
-                        <Pencil width={16} color='#FFC100' cursor='pointer' />
-                      </Link>
-                      <Trash2
-                        width={16}
-                        color='#F56565'
-                        onClick={() => deleteData(item._id, item.title)}
-                        cursor='pointer'
-                      />
-                      {/* <EyeOff width={16} color='#20B2AA' cursor='pointer' /> */}
+                      <Dialog.Root>
+                        <Dialog.Trigger>
+                          <Pencil width={16} color='#FFC100' cursor='pointer' />
+                        </Dialog.Trigger>
+
+                        <Dialog.Content style={{ maxWidth: 1100 }}>
+                          <Dialog.Title>Edit Dialect</Dialog.Title>
+                          <Dialog.Description size='2' mb='4'>
+                            Make changes to your dialect.
+                          </Dialog.Description>
+                          <div className='w-full flex items-center justify-center'>
+                            <EditDialectPage dialect={item} />
+                          </div>
+                        </Dialog.Content>
+                      </Dialog.Root>
+
+                      <AlertDialog.Root>
+                        <AlertDialog.Trigger>
+                          <Trash2 width={16} color='#F56565' cursor='pointer' />
+                        </AlertDialog.Trigger>
+                        <AlertDialog.Content style={{ maxWidth: 450 }}>
+                          <AlertDialog.Title>Revoke access</AlertDialog.Title>
+                          <AlertDialog.Description size='2'>
+                            Are you sure? This application will no longer be
+                            accessible and any existing sessions will be
+                            expired.
+                          </AlertDialog.Description>
+
+                          <Flex gap='3' mt='4' justify='end'>
+                            <AlertDialog.Cancel>
+                              <Button variant='soft' color='gray'>
+                                Cancel
+                              </Button>
+                            </AlertDialog.Cancel>
+                            <AlertDialog.Action>
+                              <Button
+                                variant='solid'
+                                color='red'
+                                onClick={() => deleteData(item._id, item.title)}
+                              >
+                                Delete
+                              </Button>
+                            </AlertDialog.Action>
+                          </Flex>
+                        </AlertDialog.Content>
+                      </AlertDialog.Root>
                     </div>
                   </Table.Cell>
                 </Table.Row>
@@ -85,9 +137,23 @@ const EditDialect = () => {
           </Table.Body>
         </Table.Root>
       </div>
-      {/* <div>
-        <Button title='Save' type='submit' onClick={() => {}} />
-      </div> */}
+      <div>
+        <Dialog.Root>
+          <Dialog.Trigger>
+            <ButtonUI title={t('control.dialect.addButton')} />
+          </Dialog.Trigger>
+
+          <Dialog.Content style={{ maxWidth: 1100 }}>
+            <Dialog.Title>Add Dialect</Dialog.Title>
+            <Dialog.Description size='2' mb='4'>
+              Make your dialect.
+            </Dialog.Description>
+            <div className='w-full flex items-center justify-center'>
+              <AddDialect />
+            </div>
+          </Dialog.Content>
+        </Dialog.Root>
+      </div>
     </motion.div>
   );
 };
